Adding a CLI for Jinja2
Allows compiling templates via the command line by taking data input
from either json, yaml, ini, or a querystring source.

Data can be read in from a file or piped in through stdin.

See __doc__ for examples on usage.

I am 100% open to critiques or adjustments that need to be made to
comply with library conformity. :)