Add support for preserving context when importing
Contrary to the documentation on the subject (see Import Context
Behavior), the context's *parent* is currently passed to imported /
included templates. This has the unintended side effect of variables
defined in the template being inaccessible inside includes that are
inside blocks (see #352 and #603).

This commit adds the "keep context" option to imports and includes,
which tells the template engine to pass the current context instead
of the parent context to the imported / included templates:

  {% from "functions.html" import my_function keep context %}
  {% import "functions.html" keep context %}
  {% include "partial.html" keep context %}