Fix parsing of the {% from %} block
There was an issue in the parser before where it was not properly
hanlding variable tuples within the parameters. Additionally, it was
not validating that "import" came after the template name, or that
when it did that any variables were actually specified. Both of those
are now being enforced properly.