Use the currently handled exception as UndefinedError.__cause__
Add ``Undefined._undefined_cause`` attribute, which is used
as the default ``__cause__`` for exceptions raised by that
``Undefined``.
When an ``Undefined`` is created, automatically set this
attribute to the currently handled exception (or ``None``).
This mirrors Python's own exception chaining.

Also, restructure `Environment.getitem` and `Environment.getattr`
uses exception chaning to preserve the exceptions they encounter.